feat: refactored the file structure of player to separate files into categories
