+++ apps/client/src/components/PlayerView.tsx
@@ -4,15 +4,15 @@ import { socket, API_URL } from "../socket";
 import { Clock, LogOut } from "lucide-react";
 import { useTranslation } from "react-i18next";
 import { LanguageSwitcher } from "./LanguageSwitcher";
-import { useCorrectTheErrorPartialScore } from "./player/useCorrectTheErrorPartialScore";
+import { useCorrectTheErrorPartialScore } from "./player/questions/correctTheError/useCorrectTheErrorPartialScore";
 import type { Competition, FillInTheBlanksContent, MatchingContent, AnswerContent, ChronologyContent, CorrectTheErrorContent, TrueFalseContent, CorrectTheErrorAnswer, MultipleChoiceContent } from "@quizco/shared";
-import { getHydratedPlayerAnswerState } from "./player/playerAnswerSync";
-import { CompetitionSelector } from "./player/CompetitionSelector";
-import { TeamJoinForm } from "./player/TeamJoinForm";
-import { WaitingPhase, RoundTransitionPhase, LeaderboardPhase } from "./player/SimplePhases";
-import { QuestionActivePhase } from "./player/QuestionActivePhase";
-import { QuestionPreviewPhase } from "./player/QuestionPreviewPhase";
-import { RevealAnswerPhase } from "./player/RevealAnswerPhase";
+import { getHydratedPlayerAnswerState } from "./player/utils/playerAnswerSync";
+import { CompetitionSelector } from "./player/lobby/CompetitionSelector";
+import { TeamJoinForm } from "./player/lobby/TeamJoinForm";
+import { WaitingPhase, RoundTransitionPhase, LeaderboardPhase } from "./player/phases/SimplePhases";
+import { QuestionActivePhase } from "./player/phases/QuestionActivePhase";
+import { QuestionPreviewPhase } from "./player/phases/QuestionPreviewPhase";
+import { RevealAnswerPhase } from "./player/phases/RevealAnswerPhase";
 
 const TEAM_ID_KEY = "quizco_team_id";
 const TEAM_NAME_KEY = "quizco_team_name";
