+++ apps/client/src/components/player/questions/DefaultReveal.tsx
