+++ apps/client/src/components/player/questions/MultipleChoiceReveal.tsx
