+++ apps/client/src/components/player/questions/correctTheError/CorrectTheErrorReveal.tsx
