+++ apps/client/src/components/player/questions/correctTheError/useCorrectTheErrorPartialScore.ts
